Extract utility `as<T>()` function for Ptr conversion

+++ YetCoreLibrary/Include/Ptr.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include <cstdint>
+#include "Primitives.h"
 
-using Ptr = uintptr_t;
+using Ptr = UInt;
